Add QuizGame component with dummy data
The question data is held inside a
DUMMY_QUESTION array which is stored inside a new
QuestionsContext. I created a new ContextProviders
component that passes the questionsContext and
SettingsContext to the app component. I also added
the QuizGame component which displays the active
Question and stores the answers in the userAnswers
state.