Fix gameDuration and QuizResults link bugs
The game duration is now stopped when
quizIsCompleted = true and is reset after the quiz
results are sent to the backend. The link in the
quiz results component now points to the correct
high scores data page according to the difficulty
of the last quiz game. Additionally, the NavLinks
for the Home navigation and High Scores navigation
now have active styling.